FWV0.7.8: Add Game of Snake!
Sub-menus updated to only disable SET button. Keep + and - to select from the list.
Port in Game of Snake!
Add W and L for win and lose in game snake.
Add time-out from game mode if the cores are not changed or game is not re-started within 30s. Re-use/adapt MenuTimeOut in Mode_Manager.
In Game Sub Menu use -/+ to cycle through the available choices, choose with S.